feat ariatemplates/ariatemplates#1427 Customization of widget's error messages
Documentation for feature described there: ariatemplates/ariatemplates#1427

Not all parts of the snippet have been prepared to be used, but it provides a common place to demonstrate all the aspects of the feature, mixing global and specific ones.

Below is the original description of the feature.

----

It is now possible for a user to customize the error messages used by the widgets.

There are 3 levels of error messages specifications, the two firsts being user configuration while the last is the hardcoded default.

This is resolved in this order of precedence:

- local: widget's configuration
- global: application's environment
- hardcoded: widget's resources